# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Enzora API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type CustomerType = typeof CustomerType[keyof typeof CustomerType];


export const CustomerType = {
  patient: 'patient',
  caregiver: 'caregiver',
  clinic: 'clinic',
  hospital: 'hospital',
  research: 'research',
  other: 'other',
} as const;

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  new: 'new',
  contacted: 'contacted',
  confirmed: 'confirmed',
  completed: 'completed',
  rejected: 'rejected',
} as const;

export type PaymentStatus = typeof PaymentStatus[keyof typeof PaymentStatus];


export const PaymentStatus = {
  pending: 'pending',
  awaiting_confirmation: 'awaiting_confirmation',
  paid: 'paid',
  failed: 'failed',
  refunded: 'refunded',
  cancelled: 'cancelled',
} as const;

export type PaymentMethodKey = typeof PaymentMethodKey[keyof typeof PaymentMethodKey];


export const PaymentMethodKey = {
  cash_on_delivery: 'cash_on_delivery',
  cash_on_pickup: 'cash_on_pickup',
  bank_transfer: 'bank_transfer',
  mobile_wallet: 'mobile_wallet',
  contact_us: 'contact_us',
} as const;

export type ProductSelection = typeof ProductSelection[keyof typeof ProductSelection];


export const ProductSelection = {
  bandage_pack: 'bandage_pack',
  smart_device: 'smart_device',
  complete_package: 'complete_package',
} as const;

export interface Product {
  productKey: ProductSelection;
  name: string;
  description: string;
  price?: number | null;
  currency: string;
  priceLabel?: string | null;
  displayText: string;
}

export interface AdminProduct {
  productKey: ProductSelection;
  name: string;
  description: string;
  price?: number | null;
  currency: string;
  priceLabel?: string | null;
  displayText: string;
  isActive: boolean;
}

export interface UpdateProductInput {
  /** @minLength 1 */
  name: string;
  description: string;
  price?: number | null;
  /** @minLength 1 */
  currency: string;
  priceLabel?: string | null;
  isActive: boolean;
}

export interface PaymentMethodPublic {
  methodKey: PaymentMethodKey;
  nameEn: string;
  nameAr: string;
  instructionsEn: string;
  instructionsAr: string;
}

export interface AdminPaymentMethod {
  methodKey: PaymentMethodKey;
  nameEn: string;
  nameAr: string;
  instructionsEn: string;
  instructionsAr: string;
  isActive: boolean;
}

export interface UpdatePaymentMethodInput {
  methodKey: PaymentMethodKey;
  /** @minLength 1 */
  nameEn: string;
  /** @minLength 1 */
  nameAr: string;
  instructionsEn: string;
  instructionsAr: string;
  isActive: boolean;
}

export interface CreateOrderInput {
  /** @minLength 1 */
  fullName: string;
  email: string;
  /** @minLength 3 */
  phone: string;
  /** Legacy combined country / city field */
  countryCity?: string;
  /** Country name */
  country?: string;
  /** City name */
  city?: string;
  customerType: CustomerType;
  productSelection: ProductSelection;
  /** @minimum 1 */
  quantity: number;
  message?: string;
  paymentMethod: PaymentMethodKey;
}

export interface OrderConfirmation {
  id: number;
  orderReference: string;
  status: OrderStatus;
  paymentMethod: PaymentMethodKey;
  paymentStatus: PaymentStatus;
  amountDue?: number | null;
  currency?: string;
}

export interface Order {
  id: number;
  orderReference: string;
  fullName: string;
  email: string;
  phone: string;
  countryCity: string;
  country?: string | null;
  city?: string | null;
  customerType: CustomerType;
  productSelection: ProductSelection;
  quantity: number;
  productNameSnapshot?: string | null;
  productPriceSnapshot?: number | null;
  productCurrencySnapshot?: string | null;
  totalEstimatedPrice?: number | null;
  message?: string | null;
  status: OrderStatus;
  paymentMethod?: PaymentMethodKey | null;
  paymentStatus: PaymentStatus;
  paymentNote?: string | null;
  paymentReference?: string | null;
  amountDue?: number | null;
  currency: string;
  paidAt?: string | null;
  createdAt: string;
}

export interface OrderTrackingResult {
  id: number;
  orderReference: string;
  productNameSnapshot: string | null;
  productSelection: ProductSelection;
  quantity: number;
  status: OrderStatus;
  paymentMethod: PaymentMethodKey | null;
  paymentStatus: PaymentStatus;
  amountDue?: number | null;
  currency: string;
  createdAt: string;
}

export interface UpdateOrderStatusInput {
  status: OrderStatus;
}

export interface UpdateOrderPaymentInput {
  paymentStatus?: PaymentStatus;
  paymentNote?: string | null;
  paymentReference?: string | null;
  amountDue?: number | null;
}

export interface AdminLoginInput {
  email: string;
  /** @minLength 1 */
  password: string;
}

export interface AdminSession {
  token: string;
}

export interface AdminMe {
  ok: boolean;
}

export interface SocialLinks {
  instagram: string | null;
  facebook: string | null;
  linkedin: string | null;
  tiktok: string | null;
}

export interface SocialLinksInput {
  instagram: string | null;
  facebook: string | null;
  linkedin: string | null;
  tiktok: string | null;
}

export type OrdersSummaryByStatus = {
  new: number;
  contacted: number;
  confirmed: number;
  completed: number;
  rejected: number;
};

export interface OrdersSummary {
  total: number;
  byStatus: OrdersSummaryByStatus;
  recent: Order[];
}

export type TrackOrderParams = {
ref: string;
};

export type ListOrdersParams = {
status?: OrderStatus;
search?: string;
};