# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Enzora API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: orders
    description: Public order/contact requests
  - name: admin
    description: Admin-protected order management
  - name: settings
    description: Public site settings
  - name: products
    description: Product and pricing management
  - name: payment-methods
    description: Payment method management
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Submit a public order/contact request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderConfirmation"

  /orders/track:
    get:
      operationId: trackOrder
      tags: [orders]
      summary: Track an order by reference number (public, safe fields only)
      parameters:
        - name: ref
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderTrackingResult"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /payment-methods:
    get:
      operationId: listPaymentMethods
      tags: [payment-methods]
      summary: List active payment methods (public)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PaymentMethodPublic"

  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List active products with live pricing
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  /social-links:
    get:
      operationId: getSocialLinks
      tags: [settings]
      summary: Get the public social media links
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SocialLinks"

  /admin/social-links:
    put:
      operationId: updateSocialLinks
      tags: [admin]
      summary: Update the public social media links
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SocialLinksInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SocialLinks"

  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login (password)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"

  /admin/me:
    get:
      operationId: adminMe
      tags: [admin]
      summary: Validate admin token
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminMe"

  /admin/products:
    get:
      operationId: adminListProducts
      tags: [admin]
      summary: List all products including inactive (admin)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminProduct"

  /admin/products/{product_key}:
    put:
      operationId: adminUpdateProduct
      tags: [admin]
      summary: Update a product's price and details
      parameters:
        - name: product_key
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminProduct"

  /admin/payment-methods:
    get:
      operationId: adminListPaymentMethods
      tags: [admin]
      summary: List all payment methods (admin)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminPaymentMethod"
    put:
      operationId: adminUpdatePaymentMethods
      tags: [admin]
      summary: Update payment method settings (active flag and instructions)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/UpdatePaymentMethodInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminPaymentMethod"

  /admin/orders:
    get:
      operationId: listOrders
      tags: [admin]
      summary: List orders (admin)
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/OrderStatus"
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"

  /admin/orders/summary:
    get:
      operationId: getOrdersSummary
      tags: [admin]
      summary: Aggregate counts and recent activity
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrdersSummary"

  /admin/orders/{id}:
    get:
      operationId: getOrder
      tags: [admin]
      summary: Get a single order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    patch:
      operationId: updateOrderStatus
      tags: [admin]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    delete:
      operationId: deleteOrder
      tags: [admin]
      summary: Delete an order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /admin/orders/{id}/payment:
    patch:
      operationId: updateOrderPayment
      tags: [admin]
      summary: Update order payment status and details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderPaymentInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    CustomerType:
      type: string
      enum: [patient, caregiver, clinic, hospital, research, other]

    OrderStatus:
      type: string
      enum: [new, contacted, confirmed, completed, rejected]

    PaymentStatus:
      type: string
      enum: [pending, awaiting_confirmation, paid, failed, refunded, cancelled]

    PaymentMethodKey:
      type: string
      enum: [cash_on_delivery, cash_on_pickup, bank_transfer, mobile_wallet, contact_us]

    ProductSelection:
      type: string
      enum: [bandage_pack, smart_device, complete_package]

    Product:
      type: object
      required: [productKey, name, description, currency, displayText]
      properties:
        productKey:
          $ref: "#/components/schemas/ProductSelection"
        name:
          type: string
        description:
          type: string
        price:
          type: number
          nullable: true
        currency:
          type: string
        priceLabel:
          type: string
          nullable: true
        displayText:
          type: string

    AdminProduct:
      type: object
      required: [productKey, name, description, currency, displayText, isActive]
      properties:
        productKey:
          $ref: "#/components/schemas/ProductSelection"
        name:
          type: string
        description:
          type: string
        price:
          type: number
          nullable: true
        currency:
          type: string
        priceLabel:
          type: string
          nullable: true
        displayText:
          type: string
        isActive:
          type: boolean

    UpdateProductInput:
      type: object
      required: [name, description, currency, isActive]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        price:
          type: number
          nullable: true
        currency:
          type: string
          minLength: 1
        priceLabel:
          type: string
          nullable: true
        isActive:
          type: boolean

    PaymentMethodPublic:
      type: object
      required: [methodKey, nameEn, nameAr, instructionsEn, instructionsAr]
      properties:
        methodKey:
          $ref: "#/components/schemas/PaymentMethodKey"
        nameEn:
          type: string
        nameAr:
          type: string
        instructionsEn:
          type: string
        instructionsAr:
          type: string

    AdminPaymentMethod:
      type: object
      required: [methodKey, nameEn, nameAr, instructionsEn, instructionsAr, isActive]
      properties:
        methodKey:
          $ref: "#/components/schemas/PaymentMethodKey"
        nameEn:
          type: string
        nameAr:
          type: string
        instructionsEn:
          type: string
        instructionsAr:
          type: string
        isActive:
          type: boolean

    UpdatePaymentMethodInput:
      type: object
      required: [methodKey, nameEn, nameAr, instructionsEn, instructionsAr, isActive]
      properties:
        methodKey:
          $ref: "#/components/schemas/PaymentMethodKey"
        nameEn:
          type: string
          minLength: 1
        nameAr:
          type: string
          minLength: 1
        instructionsEn:
          type: string
        instructionsAr:
          type: string
        isActive:
          type: boolean

    CreateOrderInput:
      type: object
      required:
        - fullName
        - email
        - phone
        - customerType
        - productSelection
        - quantity
        - paymentMethod
      properties:
        fullName:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        phone:
          type: string
          minLength: 3
        countryCity:
          type: string
          description: Legacy combined country / city field
        country:
          type: string
          description: Country name
        city:
          type: string
          description: City name
        customerType:
          $ref: "#/components/schemas/CustomerType"
        productSelection:
          $ref: "#/components/schemas/ProductSelection"
        quantity:
          type: integer
          minimum: 1
        message:
          type: string
        paymentMethod:
          $ref: "#/components/schemas/PaymentMethodKey"

    OrderConfirmation:
      type: object
      required: [id, orderReference, status, paymentMethod, paymentStatus]
      properties:
        id:
          type: integer
        orderReference:
          type: string
        status:
          $ref: "#/components/schemas/OrderStatus"
        paymentMethod:
          $ref: "#/components/schemas/PaymentMethodKey"
        paymentStatus:
          $ref: "#/components/schemas/PaymentStatus"
        amountDue:
          type: number
          nullable: true
        currency:
          type: string

    Order:
      type: object
      required:
        - id
        - orderReference
        - fullName
        - email
        - phone
        - countryCity
        - customerType
        - productSelection
        - quantity
        - status
        - paymentStatus
        - currency
        - createdAt
      properties:
        id:
          type: integer
        orderReference:
          type: string
        fullName:
          type: string
        email:
          type: string
        phone:
          type: string
        countryCity:
          type: string
        country:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        customerType:
          $ref: "#/components/schemas/CustomerType"
        productSelection:
          $ref: "#/components/schemas/ProductSelection"
        quantity:
          type: integer
        productNameSnapshot:
          type: string
          nullable: true
        productPriceSnapshot:
          type: number
          nullable: true
        productCurrencySnapshot:
          type: string
          nullable: true
        totalEstimatedPrice:
          type: number
          nullable: true
        message:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/OrderStatus"
        paymentMethod:
          $ref: "#/components/schemas/PaymentMethodKey"
          nullable: true
        paymentStatus:
          $ref: "#/components/schemas/PaymentStatus"
        paymentNote:
          type: string
          nullable: true
        paymentReference:
          type: string
          nullable: true
        amountDue:
          type: number
          nullable: true
        currency:
          type: string
        paidAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time

    OrderTrackingResult:
      type: object
      required:
        - id
        - orderReference
        - productNameSnapshot
        - productSelection
        - quantity
        - status
        - paymentMethod
        - paymentStatus
        - currency
        - createdAt
      properties:
        id:
          type: integer
        orderReference:
          type: string
        productNameSnapshot:
          type: string
          nullable: true
        productSelection:
          $ref: "#/components/schemas/ProductSelection"
        quantity:
          type: integer
        status:
          $ref: "#/components/schemas/OrderStatus"
        paymentMethod:
          $ref: "#/components/schemas/PaymentMethodKey"
          nullable: true
        paymentStatus:
          $ref: "#/components/schemas/PaymentStatus"
        amountDue:
          type: number
          nullable: true
        currency:
          type: string
        createdAt:
          type: string
          format: date-time

    UpdateOrderStatusInput:
      type: object
      required: [status]
      properties:
        status:
          $ref: "#/components/schemas/OrderStatus"

    UpdateOrderPaymentInput:
      type: object
      properties:
        paymentStatus:
          $ref: "#/components/schemas/PaymentStatus"
        paymentNote:
          type: string
          nullable: true
        paymentReference:
          type: string
          nullable: true
        amountDue:
          type: number
          nullable: true

    AdminLoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 1

    AdminSession:
      type: object
      required: [token]
      properties:
        token:
          type: string

    AdminMe:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean

    SocialLinks:
      type: object
      required: [instagram, facebook, linkedin, tiktok]
      properties:
        instagram:
          type: string
          nullable: true
        facebook:
          type: string
          nullable: true
        linkedin:
          type: string
          nullable: true
        tiktok:
          type: string
          nullable: true

    SocialLinksInput:
      type: object
      required: [instagram, facebook, linkedin, tiktok]
      properties:
        instagram:
          type: string
          nullable: true
        facebook:
          type: string
          nullable: true
        linkedin:
          type: string
          nullable: true
        tiktok:
          type: string
          nullable: true

    OrdersSummary:
      type: object
      required: [total, byStatus, recent]
      properties:
        total:
          type: integer
        byStatus:
          type: object
          required: [new, contacted, confirmed, completed, rejected]
          properties:
            new: { type: integer }
            contacted: { type: integer }
            confirmed: { type: integer }
            completed: { type: integer }
            rejected: { type: integer }
        recent:
          type: array
          items:
            $ref: "#/components/schemas/Order"